refactor(schema): 更新ID生成器并调整字段类型
- 将所有schema中的IDMixin替换为IdSonyFlakeMixin
- 统一将Uint64类型的ID字段修改为Int64类型- 更新mingyang-admin-common依赖版本至v0.1.0
- 调整tenant表parent_id默认值逻辑
- 规范oauth_provider表auth_style字段注释内容

+++ orm/ent/mixins/id_sonyflake.go
@@ -15,7 +15,7 @@ type IdSonyFlakeMixin struct {
 
 func (IdSonyFlakeMixin) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int64("id").Default(sonyflake.NextID()),
+		field.Uint64("id").Default(uint64(sonyflake.NextID())),
 		field.Time("created_at").
 			Immutable().
 			Default(time.Now).
